chore: use single class per file when possible on tests/ directory
